fixed some linting errors
