+++ .eslintrc.json
@@ -1,4 +1,13 @@
 {
-    "extends": "react-app"
-
+    "extends": "airbnb", 
+    "parser": "babel-eslint",
+    "parserOptions": {
+        "sourceType": "module",
+        "allowImportExportEverywhere": false,
+        "codeFrame": true
+      },
+    "rules": {
+        "strict": 0
+    }
+      
 }
