--- src/App.js
@@ -1,22 +1,0 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
-import { SearchPageComponent } from './components/searchPageComponent';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to Ashley's Github Profile Searcher!</h1>
-        </header>
-  
-        <SearchPageComponent />
-
-      </div>
-    );
-  }
-}
-
-export default App;
